requirements.txt
pandas
numpy
scikit-learn
matplotlib
joblib
fastapi
uvicorn[standard]
pyyaml